// admin/models/api.ts
export interface CityResponse {
  id: number;
  name: string;
  isCapital: boolean;
  state: StateResponse;
  createdAt: Date;
  updatedAt: Date;
}

export interface CreateCityRequest {
  name: string;
  isCapital: boolean;
  stateId: number;
}

export interface UpdateCityRequest {
  name: string;
  isCapital: boolean;
  stateId: number;
}

export interface Owner extends AbstractEntity {
  name: string;
  email: string;
  city: City;
  state: State;
  createdAt: Date;
  updatedAt: Date;
}

export interface CreateOwnerRequest {
  name: string;
  email: string;
  stateId: number;
  cityId: number;
}

export interface OwnerResponse {
  id: number;
  name: string;
  email: string;
  city: CityResponse;
  state: StateResponse;
  createdAt: Date;
  updatedAt: Date;
}

export interface UpdateOwnerRequest {
  name: string;
  stateId: number;
  cityId: number;
}
// ...
export interface StateResponse {
  id: number;
  name: string;
  uf: string;
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface PagedResponse<T> {
  page: number;
  size: number;
  totalElements: number;
  totalPages: number;
  data: T[];
}